refactor(tests): finish shared fixture consolidation
