+++ src/test-kernel/agent/modelHandlers/ModelHandlerContinuationContract.vitest.ts
@@ -8,6 +8,7 @@ import { ModelProvider } from 'llm-zoo';
 
 // Local imports - test support and agent
 import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
+import { noopTrace } from '@agent/trace';
 import {
   AgentCategory,
   AgentSettingSchema,
@@ -19,7 +20,6 @@ import { ModelHandlerGoogleInteractions } from '@agent/modelHandlers/google/mode
 import { ModelHandlerOpenAI } from '@agent/modelHandlers/openai/modelHandlerOpenAI';
 import { ModelHandlerOpenAIResponse } from '@agent/modelHandlers/openai/modelHandlerOpenAIResponse';
 import { ModelHandlerOpenRouterNative } from '@agent/modelHandlers/openrouter/modelHandlerOpenRouterNative';
-import { noopTrace } from '@agent/trace/noopTrace';
 
 // Type imports
 import type { MessageParam } from '@anthropic-ai/sdk/resources/messages';
