+++ src/test-kernel/agent/modelHandlers/ModelHandlerOpenAIEndTag.vitest.ts
@@ -5,9 +5,9 @@ import { ModelProvider } from 'llm-zoo';
 
 // Local imports - test support and agent
 import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
+import { noopTrace } from '@agent/trace';
 import { ModelHandlerOpenAI } from '@agent/modelHandlers/openai/modelHandlerOpenAI';
 import { ModelHandlerXAI } from '@agent/modelHandlers/openai/modelHandlerXAI';
-import { noopTrace } from '@agent/trace/noopTrace';
 
 const END_TAG = '</document>';
