+++ src/test-kernel/agent/modelHandlers/ModelHandlerOpenAINormalize.vitest.ts
@@ -5,10 +5,10 @@ import { ModelProvider } from 'llm-zoo';
 
 // Local imports - test support and agent
 import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
+import { noopTrace } from '@agent/trace';
 import { ModelHandlerDashScope } from '@agent/modelHandlers/openai/modelHandlerDashScope';
 import { ModelHandlerDeepSeek } from '@agent/modelHandlers/openai/modelHandlerDeepSeek';
 import { ModelHandlerKimi } from '@agent/modelHandlers/openai/modelHandlerKimi';
-import { noopTrace } from '@agent/trace/noopTrace';
 
 function createClientStub() {
   const createCalls: any[] = [];
