+++ src/test-kernel/agent/modelHandlers/OpenAICompatibleProviderParams.vitest.ts
@@ -7,11 +7,11 @@ import { ModelProvider, ReasoningEffort } from 'llm-zoo';
 
 // Local imports - test support and agent
 import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
+import { noopTrace } from '@agent/trace';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import { ModelHandlerGLM } from '@agent/modelHandlers/openai/modelHandlerGLM';
 import { ModelHandlerKimi } from '@agent/modelHandlers/openai/modelHandlerKimi';
 import { ModelHandlerXAI } from '@agent/modelHandlers/openai/modelHandlerXAI';
-import { noopTrace } from '@agent/trace/noopTrace';
 
 const NO_VISION_CAPABILITIES = Object.freeze({ supportsVision: false });
 const MOONSHOT_TEST_CONFIG = Object.freeze({
